requirements.txt
# System 
python-dotenv

# Data processing
pandas
scikit-learn
matplotlib
umap-learn
sentence-transformers

# Data connectivity
pysam
httpx
tqdm

# Scrapping
bs4
mwclient
lxml